+++ src/main/kotlin/icons/AMIIIcons.kt
@@ -20,6 +20,9 @@ object AMIIIcons {
   @JvmField
   val E1F44B = IconLoader.getIcon("/icons/emojis/1f44b.svg", javaClass)
 
+  @JvmField
+  val E1f648 = IconLoader.getIcon("/icons/emojis/1f648.svg", javaClass)
+
   val E1F60A = IconLoader.getIcon("/icons/emojis/1f60a.svg", javaClass)
   val E1F60C = IconLoader.getIcon("/icons/emojis/1f60c.svg", javaClass)
   val E1F60F = IconLoader.getIcon("/icons/emojis/1f60f.svg", javaClass)