+++ src/main/kotlin/io/unthrottled/amii/core/MIKU.kt
@@ -20,6 +20,7 @@ import io.unthrottled.amii.core.personality.emotions.EmotionalMutationActionList
 import io.unthrottled.amii.core.personality.emotions.EmotionalMutationType
 import io.unthrottled.amii.core.personality.emotions.Mood
 import io.unthrottled.amii.core.personality.emotions.MoodListener
+import io.unthrottled.amii.discreet.discreetModeService
 import io.unthrottled.amii.events.UserEvent
 import io.unthrottled.amii.events.UserEventListener
 import io.unthrottled.amii.events.UserEvents
@@ -48,7 +49,7 @@ class MIKU(private val project: Project) :
   }
 
   private var emotionCore = EmotionCore(Config.instance)
-  private var modalityProcessor = ModalityProcessor(Config.instance)
+  private var modalityProcessor = ModalityProcessor(Config.instance, project)
   private val taskPersonalityCore = TaskPersonalityCore()
   private val onDemandPersonalityCore = OnDemandPersonalityCore()
   private val alertPersonalityCore = AlertPersonalityCore()
@@ -106,6 +107,8 @@ class MIKU(private val project: Project) :
   }
 
   private fun consumeEvents(bufferedUserEvents: List<UserEvent>) {
+    if (project.discreetModeService().isDiscreetMode) return
+
     val emotionalState = emotionCore.deriveMood(bufferedUserEvents.first())
     bufferedUserEvents.forEach { userEvent -> reactToEvent(userEvent, emotionalState) }
     publishMood(emotionalState)
@@ -130,6 +133,8 @@ class MIKU(private val project: Project) :
   private fun reactToMutation(
     emotionalMutationAction: EmotionalMutationAction
   ) {
+    if (project.discreetModeService().isDiscreetMode) return
+
     if (emotionalMutationAction.type == EmotionalMutationType.RESET) {
       resetCore.processMutationEvent(emotionalMutationAction)
     }