+++ src/main/kotlin/io/unthrottled/amii/discreet/ProjectDiscreetModeListener.kt
@@ -1,0 +1,12 @@
+package io.unthrottled.amii.discreet
+
+import com.intellij.openapi.project.Project
+
+class ProjectDiscreetModeListener(private val project: Project) : DiscreetModeListener {
+  override fun modeChanged(discreetMode: DiscreetMode) {
+    when (discreetMode) {
+      DiscreetMode.ACTIVE -> project.discreetModeService().applyDiscreetMode()
+      DiscreetMode.INACTIVE -> project.discreetModeService().liftDiscreetMode()
+    }
+  }
+}