+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 # AMII Changelog
 
+## [0.12.0]
+
+### Added
+
+- `Discreet Mode`, when enabled MIKU will clear and not show _any_ anime in the IDE. Also, the mood in the status bar
+  will temporarily hide as well.
+
 ## [0.11.0]
 
 ### Added